Updated makefile so it compiles with latest toolchain and libctru.
Note that it will still not *run* without a modified libcitro3d to raise
the maximum queue length.  The default of 32 (which, as of writing, is
not changeable from user code with upstream citro3d) is *far* too low
and the first level will not even load with it set to less than 128 (256
recommended).